+ DefaultBotSettings
DefaultBotSettings that are generated on loading the world if they don't exist.
These include Json files stored in the path: "WorldSave/Finance/StockMarket/DefaultBotSettings/"
The name of the files does not matter, several files are used to classify the items.
This allows you to find certain items more quickly.
The structure of the files is always the same. {"itemID":{settings...}...}
The values in the json will be used as default values when the command "/StockMarket createDefaultBot <itemID>" is called.